+++ src/components/generics/vacation-test-forms/myVacationMockData.json
@@ -7,7 +7,12 @@
         "days": 2,
         "startDate": "01.04.2023",
         "endDate": "02.04.2023",
-        "status": "ACCEPTED"
+        "remainingDays": 25,
+        "status": "ACCEPTED",
+        "created": "2023-01-25",
+        "updated": "2023-01-30",
+        "projectManager": "Approved",
+        "humanResourcesManager": "Pending"
     },
     {
         "id": 29582526,
@@ -17,7 +22,12 @@
         "days": 2,
         "startDate": "15.04.2023",
         "endDate": "16.04.2023",
-        "status": "PENDING"
+        "remainingDays": 25,
+        "status": "PENDING",
+        "created": "2023-01-25",
+        "updated": "2023-01-30",
+        "projectManager": "Approved",
+        "humanResourcesManager": "Pending"
     },
     {
         "id": 3951226,
@@ -27,7 +37,12 @@
         "days": 10,
         "startDate": "01.06.2023",
         "endDate": "14.06.2023",
-        "status": "PENDING"
+        "remainingDays": 25,
+        "status": "PENDING",
+        "created": "2023-01-25",
+        "updated": "2023-01-30",
+        "projectManager": "Approved",
+        "humanResourcesManager": "Pending"
     },
     {
         "id": 4958226,
@@ -37,6 +52,11 @@
         "days": 15,
         "startDate": "01.02.2023",
         "endDate": "21.02.2023",
-        "status": "DECLINED"
+        "remainingDays": 25,
+        "status": "DECLINED",
+        "created": "2023-01-25",
+        "updated": "2023-01-30",
+        "projectManager": "Approved",
+        "humanResourcesManager": "Pending"
     }
 ]
